Implement basic splatting in rt

+++ src/renderer.hpp
@@ -8,7 +8,7 @@ namespace bu {
 struct renderer 
 {
 	virtual ~renderer() = default;
-	virtual void draw(const bu::scene &scene, const bu::camera &camera, const glm::vec2 &viewwport_size) = 0;
+	virtual void draw(const bu::scene &scene, const bu::camera &camera, const glm::ivec2 &viewwport_size) = 0;
 };
 
 }
