+++ src/renderers/albedo/albedo.hpp
@@ -22,7 +22,7 @@ class albedo_renderer : public bu::renderer
 {
 public:
 	albedo_renderer(std::shared_ptr<albedo_context> &context);
-	void draw(const bu::scene &scene, const bu::camera &camera, const glm::vec2 &viewport_size) override;
+	void draw(const bu::scene &scene, const bu::camera &camera, const glm::ivec2 &viewport_size) override;
 
 private:
 	std::shared_ptr<albedo_context> m_context;
